Conflicts:
	src/main/java/fr/miage/projetagent/entity/EnvoiVaccin.java

+++ src/main/java/fr/miage/projetagent/agent/Priority.java
@@ -7,7 +7,7 @@ public class Priority {
     //date where a patient will die
     private Date date;
     //volume of vaccine we already have for this disease and this date
-    private int volume;
+    private double volume;
     //number of people to cure, minus the number of vaccine we already have for this disease and this date
     private int nombre;
     private String pays;
@@ -32,11 +32,11 @@ public class Priority {
         this.date = date;
     }
 
-    public int getVolume() {
+    public double getVolume() {
         return volume;
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
